+++ sketch-compose-core/src/main/kotlin/com/github/panpf/sketch/compose/AsyncImageState.kt
@@ -39,6 +39,7 @@ import com.github.panpf.sketch.compose.internal.AsyncImageScaleDecider
 import com.github.panpf.sketch.compose.internal.AsyncImageSizeResolver
 import com.github.panpf.sketch.compose.internal.CrossfadePainter
 import com.github.panpf.sketch.compose.internal.findLeafChildPainter
+import com.github.panpf.sketch.compose.internal.isEmpty
 import com.github.panpf.sketch.compose.internal.toScale
 import com.github.panpf.sketch.request.DisplayRequest
 import com.github.panpf.sketch.request.DisplayResult
@@ -120,6 +121,13 @@ class AsyncImageState internal constructor(
         private set
 
     fun setSize(size: IntSize) {
+        val oldSize = this.size
+        if (oldSize != null && oldSize.isEmpty()) {
+            // The width or height of oldSize is 0, which means that the width or height of AsyncImage is wrap content.
+            // Then only the size set for the first time can be used as the size of the image request,
+            // because the first size can accurately represent the width and height of AsyncImage.
+            return
+        }
         this.size = size
         this.sizeResolver.sizeState.value = size
     }