+++ sketch-core/src/androidTest/kotlin/com/github/panpf/sketch/core/test/decode/internal/DecodeUtilsTest.kt
@@ -565,28 +565,28 @@ class DecodeUtilsTest {
         )
 
 
-        val maxSize = OpenGLTextureHelper.maxSize ?: 0
-        val expected = when {
-            maxSize <= 4096 -> 32
-            maxSize <= 8192 -> 16
-            else -> 4
-        }
-        Assert.assertEquals(
-            expected,
-            calculateSampleSize(
-                imageSize = Size(30000, 750),
-                targetSize = Size(1080, 1920),
-                smallerSizeMode = false
-            )
-        )
-        Assert.assertEquals(
-            32,
-            calculateSampleSize(
-                imageSize = Size(30000, 750),
-                targetSize = Size(1080, 1920),
-                smallerSizeMode = true
-            )
-        )
+//        val maxSize = OpenGLTextureHelper.maxSize ?: 0
+//        val expected = when {
+//            maxSize <= 4096 -> 32
+//            maxSize <= 8192 -> 16
+//            else -> 4
+//        }
+//        Assert.assertEquals(
+//            expected,
+//            calculateSampleSize(
+//                imageSize = Size(30000, 750),
+//                targetSize = Size(1080, 1920),
+//                smallerSizeMode = false
+//            )
+//        )
+//        Assert.assertEquals(
+//            32,
+//            calculateSampleSize(
+//                imageSize = Size(30000, 750),
+//                targetSize = Size(1080, 1920),
+//                smallerSizeMode = true
+//            )
+//        )
     }
 
     @Test
@@ -835,28 +835,28 @@ class DecodeUtilsTest {
             )
         )
 
-        val maxSize = OpenGLTextureHelper.maxSize ?: 0
-        val expected = when {
-            maxSize <= 4096 -> 32
-            maxSize <= 8192 -> 16
-            else -> 4
-        }
-        Assert.assertEquals(
-            expected,
-            calculateSampleSizeForRegion(
-                regionSize = Size(30000, 750),
-                targetSize = Size(1080, 1920),
-                smallerSizeMode = false
-            )
-        )
-        Assert.assertEquals(
-            32,
-            calculateSampleSizeForRegion(
-                regionSize = Size(30000, 750),
-                targetSize = Size(1080, 1920),
-                smallerSizeMode = true
-            )
-        )
+//        val maxSize = OpenGLTextureHelper.maxSize ?: 0
+//        val expected = when {
+//            maxSize <= 4096 -> 32
+//            maxSize <= 8192 -> 16
+//            else -> 4
+//        }
+//        Assert.assertEquals(
+//            expected,
+//            calculateSampleSizeForRegion(
+//                regionSize = Size(30000, 750),
+//                targetSize = Size(1080, 1920),
+//                smallerSizeMode = false
+//            )
+//        )
+//        Assert.assertEquals(
+//            32,
+//            calculateSampleSizeForRegion(
+//                regionSize = Size(30000, 750),
+//                targetSize = Size(1080, 1920),
+//                smallerSizeMode = true
+//            )
+//        )
     }
 
     @Test