+++ sketch-core/src/main/kotlin/com/github/panpf/sketch/util/Size.kt
@@ -81,3 +81,13 @@ fun Size.isSameAspectRatio(other: Size, delta: Float = 0f): Boolean {
     }
     return false
 }
+
+fun Size.coerceAtLeast(minimumValue: Size): Size = Size(
+    width = width.coerceAtLeast(minimumValue.width),
+    height = height.coerceAtLeast(minimumValue.height),
+)
+
+fun Size.coerceAtMost(maximumValue: Size): Size = Size(
+    width = width.coerceAtMost(maximumValue.width),
+    height = height.coerceAtMost(maximumValue.height),
+)